+++ processor/src/operations/mod.rs
@@ -10,9 +10,9 @@ mod fri_ops;
 mod horner_ops;
 mod io_ops;
 mod stack_ops;
-mod sys_ops;
+pub(crate) mod sys_ops;
 mod u32_ops;
-mod utils;
+pub(crate) mod utils;
 
 #[cfg(test)]
 use super::Kernel;