+++ processor/src/stack/mod.rs
@@ -354,7 +354,7 @@ impl Stack {
 
     /// Increments the clock cycle.
     pub fn advance_clock(&mut self) {
-        self.clk += 1;
+        self.clk += 1_u32;
         self.overflow.advance_clock();
     }