+++ processor/src/stack/overflow.rs
@@ -204,7 +204,7 @@ impl OverflowTable {
 
     /// Increments the clock cycle.
     pub fn advance_clock(&mut self) {
-        self.clk += 1;
+        self.clk += 1_u32;
     }
 
     // HELPERS