+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - Remove `FALCON_SIG_TO_STACK` event (#1703).
 - Prevent `U64Div` event from crashing processor (#1710).
 - [BREAKING] Improve processor errors for memory and calls (#1717)
+- Implement a new fast processor that doesn't generate a trace (#1668)
 - `ProcessState::get_stack_state()` now only returns the state of the active context (#1753)
 
 ## 0.12.0 (2025-01-22)